Replace _.each with for...in call because of length property in sensors config for IOT
As stated by the docs "Collection functions work on arrays, objects, and array-like objects such as arguments, NodeList and similar. But it works by duck-typing, so avoid passing objects with a numeric length property." the each function will not work properly when an object has an attribute with the name length. As the IOT Configuration Sensors has a property length the compose failed and returned empty objects